updated github actions for windows

+++ typescript/build.ps1
@@ -1,4 +1,4 @@
-Remove-Item sea-prep.blob
+if (Test-Path sea-prep.blob) { Remove-Item sea-prep.blob }
 npm run esbuild
 node --experimental-sea-config sea-config.json
 node -e "require('fs').copyFileSync(process.execPath, 'athan.exe')"
